// src/shared/api/openapi/models/PostsPerMonth.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Lazy.Blog.Api | v1
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PostsPerMonth
 */
export interface PostsPerMonth {
    /**
     * 
     * @type {number}
     * @memberof PostsPerMonth
     */
    year: number;
    /**
     * 
     * @type {number}
     * @memberof PostsPerMonth
     */
    month: number;
    /**
     * 
     * @type {number}
     * @memberof PostsPerMonth
     */
    count: number;
}

/**
 * Check if a given object implements the PostsPerMonth interface.
 */
export function instanceOfPostsPerMonth(value: object): value is PostsPerMonth {
    if (!('year' in value) || value['year'] === undefined) return false;
    if (!('month' in value) || value['month'] === undefined) return false;
    if (!('count' in value) || value['count'] === undefined) return false;
    return true;
}

export function PostsPerMonthFromJSON(json: any): PostsPerMonth {
    return PostsPerMonthFromJSONTyped(json, false);
}

export function PostsPerMonthFromJSONTyped(json: any, ignoreDiscriminator: boolean): PostsPerMonth {
    if (json == null) {
        return json;
    }
    return {
        
        'year': json['year'],
        'month': json['month'],
        'count': json['count'],
    };
}

export function PostsPerMonthToJSON(json: any): PostsPerMonth {
    return PostsPerMonthToJSONTyped(json, false);
}

export function PostsPerMonthToJSONTyped(value?: PostsPerMonth | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'year': value['year'],
        'month': value['month'],
        'count': value['count'],
    };
}